fixed makebuckets to allow for selecting which buckets to recreate
git-svn-id: https://source.sunlightlabs.com/sunlightapi/trunk@119 b95ab466-21ef-4d7f-887f-4deb394a2fdb

+++ api/management/commands/makebuckets.py
@@ -6,9 +6,8 @@ from sunlightapi.legislators.models import Legislator, LegislatorBucket
 from sunlightapi.lobbyists.models import Lobbyist, LobbyistBucket
 
 def create_buckets(person, bucket_type):
-    
+
     # first and last initial
-    
     if person.firstname:
         first_initial = person.firstname[0] 
         last_initial = person.lastname[0]
@@ -30,19 +29,28 @@ def create_buckets(person, bucket_type):
         pass
 
 class Command(BaseCommand):
-    option_list = BaseCommand.option_list
+    option_list = BaseCommand.option_list + (
+        make_option('--legislators', action='store_true', dest='legislators',
+                    default=False, help='refresh legislator buckets'),
+        make_option('--lobbyists', action='store_true', dest='lobbyists',
+                    default=False, help='refresh lobbyist buckets'),)
+
     help = "Create the buckets for loose matching of people's names"
     args = ""
     requires_model_validation = False
 
     def handle(self, *fname, **options):
+        legislators = options.get('legislators', False)
+        lobbyists = options.get('lobbyists', False)
         # fill all the buckets
-        LegislatorBucket.objects.delete()
-        for leg in Legislator.objects.all():
-            print leg
-            create_buckets(leg, LegislatorBucket)
+        if legislators:
+            LegislatorBucket.objects.all().delete()
+            for leg in Legislator.objects.all():
+                print leg
+                create_buckets(leg, LegislatorBucket)
 
-        LobbyistBucket.objects.delete()
-        for lobbyist in Lobbyist.objects.all():
-            print lobbyist
-            create_buckets(lobbyist, LobbyistBucket)
+        if lobbyists:
+            LobbyistBucket.objects.delete()
+            for lobbyist in Lobbyist.objects.all():
+                print lobbyist
+                create_buckets(lobbyist, LobbyistBucket)
